Завершает ДЗ(1)
