+++ js/form.js
@@ -40,4 +40,39 @@ const closeUserModal = () => {
   resetEffects();
 };
 
-export {setFormHandlers, closeUserModal, formUpload};
+const success = document.querySelector('#success')
+  .content
+  .querySelector('.success');
+
+const err = document.querySelector('#error')
+  .content
+  .querySelector('.error');
+
+const onPopupEscKeydownSuccess = (evt) => {
+  if (evt.key === 'Escape') {
+    evt.preventDefault();
+    success.remove();
+    err.remove();
+    document.removeEventListener('keydown', onPopupEscKeydownSuccess);
+  }
+};
+
+const createSuccess = () => {
+  document.body.appendChild(success);
+  document.addEventListener('keydown', onPopupEscKeydownSuccess);
+  success.addEventListener('click', () => {
+    success.remove();
+    document.removeEventListener('keydown', onPopupEscKeydownSuccess);
+  });
+};
+
+const createErr = () => {
+  document.body.appendChild(err);
+  document.addEventListener('keydown', onPopupEscKeydownSuccess);
+  err.addEventListener('click', () => {
+    err.remove();
+    document.removeEventListener('keydown', onPopupEscKeydownSuccess);
+  });
+};
+
+export {setFormHandlers, closeUserModal, formUpload, createSuccess, createErr};
