+++ js/main.js
@@ -1,5 +1,5 @@
 import {createObjects} from './create-object.js';
-import {setFormHandlers, closeUserModal} from './form.js';
+import {setFormHandlers, closeUserModal, createErr} from './form.js';
 import {getPhotos, setUserFormSubmit} from './api.js';
 import {createSimilarPhotosAbove} from './thumbs.js';
 
@@ -7,4 +7,4 @@ createObjects();
 setFormHandlers();
 getPhotos(createSimilarPhotosAbove);
 
-setUserFormSubmit(closeUserModal);
+setUserFormSubmit(closeUserModal, createErr);
